Upgrading Umbraco version to latest.
Adding in support for custom alias. This is to give control over the naming of these as core umbraco properties require specific alias to work. This could be achieved naming the property the same as the alias but then coupled to naming convention of property on model.

Added in Is Container as this is required for project whereas folders are used for structuring document types.